LazyCall.jl

LazyCall allows you to store a function along with its arguments for later use. This is particularly useful in functional programming.

Methods on LazyCalls are currently defined for one Base function: broadcast. Use lazy_call_methods to easily define more.

LazyCall.Call

LazyCall.bit_not

LazyCall.collect_call

LazyCall.lazy_call_methods

LazyCall.push

LazyCall.unshift

LazyCall.unweave

LazyCall.@lazy_call_methods

LazyCall.@unweave

LazyCall.Call — Type.

immutable Call

Will store positional and keyword arguments. Create with collect_call or unweave.

You can include positional arguments, keyword arguments, or both.

julia> using LazyCall

julia> collect_call(1, 2)
1, 2

julia> collect_call(a = 1, b = 2)
; a = 1, b = 2

julia> collect_call(1, 2, a = 1, b = 2)
1, 2; a = 1, b = 2

They are equal if all of their elements are equal and the positional arguments are in the right order. Keyword order doesn't matter.

julia> using LazyCall

julia> collect_call(1, 2, a = 1, b = 2) == collect_call(1, 2, b = 2, a = 1)
true

You can merge two Calls.

julia> using LazyCall, ChainRecursive

julia> @chain begin
           merge(
               collect_call(1, a = 2, b = 3),
               collect_call(4, a = 5, c = 6)
            )
            _ == collect_call(1, 4; a = 5, b = 3, c = 6)
       end
true

You can push in new arguments.

julia> using LazyCall, ChainRecursive

julia> @chain begin
           collect_call(1, a = 2, b = 3)
           push(_, 4, a = 5, c = 6)
           _ == collect_call(1, 4; a = 5, b = 3, c = 6)
       end
true

You can unshift in new arguments.

julia> using LazyCall, ChainRecursive

julia> @chain begin
           collect_call(1, a = 2, b = 3)
           unshift(_, 4, a = 5, c = 6)
           _ == collect_call(4, 1; a = 2, b = 3, c = 6)
       end
true

When you run a Call, it calls the first positional argument on the rest of the positional and keyword arguments.

julia> using LazyCall

julia> collect_call(vcat, 1, 2) |> run
2-element Array{Int64,1}:
 1
 2

source

LazyCall.bit_not — Function.

bit_not

Alias for ~ for use within @unweave.

examples

julia> using LazyCall

julia> bit_not(1) == ~1
true

source

LazyCall.collect_call — Method.

collect_call(positional...; keyword...)

Collect a Call.

julia> using LazyCall

julia> collect_call(1, 2, a = 1, b = 2)
1, 2; a = 1, b = 2

source

LazyCall.lazy_call_methods — Method.

@lazy_call_methods a_module fs...

For each function in fs in a_module, create a method that takes a Call. This method will make assumptions about argument order. Positional arguments from the outer call will be inserted between the first (typically a function) and second positional arguments of the inner [Call]. This makes sense in many cases. For more complicated cases, like mapreducedim, Call methods can be defined by hand.

julia> using LazyCall, ChainRecursive

julia> @lazy_call_methods Base Generator mapreduce

julia> version_1 = @chain begin
           [-2, 0, -2, 0]
           @unweave ~_ + 1 > 0
           Base.Generator(_)
           sum(_)
        end
2

julia> version_2 = @chain begin
           [-2, 0, -2, 0]
           @unweave ~_ + 1 > 0
           mapreduce(_, +)
       end
2

julia> version_1 == version_2
true

source

LazyCall.push — Method.

push

Same as push! but will leave arguments unchanged.

source

LazyCall.unshift — Method.

unshift

Same as unshift! but will leave arguments unchanged.

source

LazyCall.@lazy_call_methods — Macro.

See documentation of lazy_call_methods

source

LazyCall.@unweave — Macro.

See documentation of unweave

source

LazyCall.unweave — Method.

@unweave e

Interprets e as a function with its arguments wrapped in tildas and woven into it. Will return a Call, with an anonymous function as the first positional argument.

Both variables and expressions can be woven in.

julia> using LazyCall, ChainRecursive

julia> A = [1, 2];

julia> @chain begin
           @unweave vcat(~A, ~[3, 4] )
           broadcast(_)
           _ == broadcast(vcat, A, [3, 4] )
       end
true

Variables need only be marked once as arguments.

julia> using LazyCall, ChainRecursive

julia> A = [1, 2];

julia> @chain begin
           @unweave vcat(~A, A)
           broadcast(_)
           _ == map(vcat, A, A)
       end
true

No more than one splatted positional argument can be woven in.

julia> using LazyCall, ChainRecursive

julia> A = [1, 2], [3, 4];

julia> B = [5, 6], [7, 8];

julia> @chain begin
           @unweave vcat( ~(A...) )
           broadcast(_)
           _ == broadcast(vcat, A...)
       end
true

julia> @unweave vcat( ~(A...), ~(B...) )
ERROR: syntax: invalid ... on non-final argument
[...]

No more than one splatted keyword argument can be woven in.

julia> using LazyCall, ChainRecursive

julia> keyword_arguments(; kwargs...) = kwargs;

julia> A = keyword_arguments( a = 1, b = 2);

julia> B = keyword_arguments( c = 3, d = 4);

julia> @chain begin
           @unweave keyword_arguments(; ~(A...) )
           run(_)
           _ == keyword_arguments(; A...)
       end
true

julia> @unweave keyword_arguments(; ~( A...), ~(B...) )
ERROR: Can only weave in one (set of) parameters
[...]

With no woven arguments, Call will only contain a dummy anonymous function.

julia> using LazyCall

julia> run(@unweave 1)
1

source